+++ packages/eyes-storybook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add `sendDom` config param
 
 ## 3.22.11 - 2021/8/16